Booking Refactoring R6 Implementar el patron mapaeador en los repositorios
